sre-agent/requirements.txt
azure-identity>=1.15.0
azure-monitor-query>=1.3.0
azure-servicebus>=7.12.0
openai>=1.14.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
